Code Review
The major change i made was to display the instructions for the program
in a texture. Since very specific buttons did very specific things, i
thought that it would be better to show what buttons did what while the
program was running instead of having to look through the code. I also
changed the reverse key to 'r' just to make things a little cleaner. The
displayed instructions can be toggled with '?' as well. Next (i'm not
sure if this is an improvement or just a change) I added transparency,
both to help the instructions be readable, and give another layer of
interest into the product. I ran out of time, but i was trying to devise
a way to improve the 'send to the back on a mouse click' method.
Currently, if you click on an area where two circles overlap, the
program might not handle the input as intended. The circle that is
"ontop" will not be sent to the back. I might be thinking about that
wrong, but if i had more time i would look further into that.